Added first framework for BAMI simulations
Note that this is the very first version of the BAMI simulation framework. This code is the combination of my simulation code that I used to evaluate TrustChain and AnyDex. Many features are missing, including scenario files. Also, the code is unoptimized and simulations might require more memory than expected, especially with more than 100 peers.